+++ integration-tests/tests/abi/json_schema.rs
