+++ integration-tests/tests/lib.rs
@@ -1,3 +1,3 @@
+mod abi;
 mod build;
 mod cargo;
-mod tests;
